# server/openapi.yaml
openapi: 3.0.3
info:
  title: honly API
  version: 0.1.0
  description: API dla panelu salonu, klienta i publicznej rezerwacji.
servers:
  - url: https://honly.pl
    description: Production
  - url: http://localhost:4000
    description: Local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthError:
      type: object
      properties:
        error:
          type: string
          enum:
            - invalid_payload
            - invalid_credentials
            - account_inactive
            - client_profile_missing
            - token_invalid
            - salon_not_linked
            - unauthorized
            - forbidden
            - email_taken
            - internal_error
            - profile_not_found
            - account_not_found
            - invalid_current_password
      required: [error]
...
    ClientAuthResponse:
      type: object
      properties:
        token: { type: string }
        clientId: { type: string }
        salonId: { type: string }
        salons:
          type: array
          items:
            $ref: '#/components/schemas/ClientSalon'
    ClientSalon:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        clientId: { type: string }
        address: { type: string }
        phone: { type: string }
        hours: { type: string }
        description: { type: string }
...
paths:
...
  /api/client/login:
    post:
      summary: Login klienta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
              required: [email, password]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ClientAuthResponse' }
              examples:
                success:
                  value:
                    token: eyJhbGciOi...
                    clientId: c_123
                    salonId: s_123
                    salons:
                      - id: s_123
                        name: Studio Bella
                        slug: studio-bella
                        clientId: c_123
                        address: "ul. Kwiatowa 1, Warszawa"
                        phone: "+48 123 456 789"
                        hours: "Pn-Pt 9:00-18:00"
                        description: "Salon fryzjerski"
                      - id: s_456
                        name: Beauty House
                        slug: beauty-house
                        clientId: c_456
                        address: "ul. Lipowa 5, Warszawa"
                        phone: "+48 222 333 444"
                        hours: "Pn-Sb 10:00-20:00"
                        description: "Salon kosmetyczny"
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                invalid_payload:
                  value: { error: invalid_payload }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                invalid_credentials:
                  value: { error: invalid_credentials }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                account_inactive:
                  value: { error: account_inactive }
        '500':
          description: Internal error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                internal_error:
                  value: { error: internal_error }
  /api/client/salons:
    get:
      summary: Lista salonów klienta
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  salons:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClientSalon'
                  activeSalonId:
                    type: string
              examples:
                success:
                  value:
                    activeSalonId: s_123
                    salons:
                      - id: s_123
                        name: Studio Bella
                        slug: studio-bella
                        clientId: c_123
                        address: "ul. Kwiatowa 1, Warszawa"
                        phone: "+48 123 456 789"
                        hours: "Pn-Pt 9:00-18:00"
                        description: "Salon fryzjerski"
                      - id: s_456
                        name: Beauty House
                        slug: beauty-house
                        clientId: c_456
                        address: "ul. Lipowa 5, Warszawa"
                        phone: "+48 222 333 444"
                        hours: "Pn-Sb 10:00-20:00"
                        description: "Salon kosmetyczny"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                unauthorized:
                  value: { error: unauthorized }
        '404':
          description: Account not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                account_not_found:
                  value: { error: account_not_found }
  /api/client/switch-salon:
    post:
      summary: Zmiana aktywnego salonu klienta
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                salonId: { type: string }
              required: [salonId]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  clientId: { type: string }
                  salonId: { type: string }
              examples:
                success:
                  value:
                    token: eyJhbGciOi...
                    clientId: c_456
                    salonId: s_456
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                invalid_payload:
                  value: { error: invalid_payload }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                unauthorized:
                  value: { error: unauthorized }
        '404':
          description: Salon not linked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                salon_not_linked:
                  value: { error: salon_not_linked }
  /api/client/salons/attach:
    post:
      summary: Dodanie salonu klienta po tokenie
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
              required: [token]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  salons:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClientSalon'
              examples:
                success:
                  value:
                    ok: true
                    salons:
                      - id: s_123
                        name: Studio Bella
                        slug: studio-bella
                        clientId: c_123
                        address: "ul. Kwiatowa 1, Warszawa"
                        phone: "+48 123 456 789"
                        hours: "Pn-Pt 9:00-18:00"
                        description: "Salon fryzjerski"
                      - id: s_456
                        name: Beauty House
                        slug: beauty-house
                        clientId: c_456
                        address: "ul. Lipowa 5, Warszawa"
                        phone: "+48 222 333 444"
                        hours: "Pn-Sb 10:00-20:00"
                        description: "Salon kosmetyczny"
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                invalid_payload:
                  value: { error: invalid_payload }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                unauthorized:
                  value: { error: unauthorized }
        '404':
          description: Token invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                token_invalid:
                  value: { error: token_invalid }
  /api/client/password-reset:
    post:
      summary: Reset hasła klienta (request)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
              required: [email]
      responses:
        '200':
          description: OK
        '400':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                invalid_payload:
                  value: { error: invalid_payload }
  /api/client/password-reset/confirm:
    post:
      summary: Reset hasła klienta (confirm)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                newPassword: { type: string }
              required: [token, newPassword]
      responses:
        '200':
          description: OK
        '400':
          description: Invalid payload / token invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthError' }
              examples:
                invalid_payload:
                  value: { error: invalid_payload }
                token_invalid:
                  value: { error: token_invalid }
  /api/client/me:
    get:
      summary: Profil klienta
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
    put:
      summary: Aktualizacja profilu klienta
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
                email: { type: string }
              required: [name, phone]
      responses:
        '200':
          description: OK
  /api/client/password:
    put:
      summary: Zmiana hasła klienta
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword: { type: string }
                newPassword: { type: string }
              required: [currentPassword, newPassword]
      responses:
        '200':
          description: OK
  /api/client/appointments:
    get:
      summary: Wizyty klienta (z cancelToken)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: OK
...
  /api/salon/clients/{id}/appointments:
    get:
      summary: Wizyty klienta w salonie
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...